TDD 리팩터링: Step3 - 로그인 리팩터링 (#51)
* test(auth): Interceptor기반 JwtToken을 사용한 인증 기능 테스트 코드 작성

* feature(auth): interceptor 기반 JwtToken 인증 기능 구현

* feature(auth): WebMvcConfig에 TokenSecurityContextPersistenceInterceptor 등록

* fix(auth): Token이 유효하지 않은 경우 아무 것도 하지 않도록 변경

* test(auth): TokenAuthenticationInterceptorTest의 단위테스트 추가

* test(auth): 테스트 코드의 흐름을 읽기 쉽게하기 위해 given/when/then 주석 추가

* test(auth): TokenSecurityContextPersistenceInterceptor.afterCompletion 테스트 코드 추가

* feature(member): Token을 사용한 회원 정보 수정/삭제 기능 구현

* test(member): Token을 사용한 회원 정보 수정/삭제 기능 인수테스트 작성

* feature(auth): 인증 정보를 파라미터로 가져오기 위한 HandlerMethodArgumentResolver 구현

* test(auth): 깨지는 테스트 수정

* refactor(member): @AuthenticationPrincipal를 사용하여 SecurityContextHolder에 직접 접근하는 코드 리팩토링

* refactor(auth): Session 기반 로그인 기능 제거

* refactor(auth): 로그인한 유저의 정보를 담는 LoginMember를 member에서 auth로 이동

* refactor(auth): 패키지 순환 의존성 해결

* refactor(auth): 변경하지 않은 메서드명 변경

+++ src/main/java/nextstep/subway/member/ui/MemberController.java
@@ -1,19 +1,20 @@
 package nextstep.subway.member.ui;
 
-import nextstep.subway.auth.domain.Authentication;
-import nextstep.subway.auth.infrastructure.SecurityContextHolder;
+import nextstep.subway.auth.domain.AuthenticationPrincipal;
+import nextstep.subway.auth.domain.UserDetails;
 import nextstep.subway.member.application.MemberService;
-import nextstep.subway.member.domain.LoginMember;
 import nextstep.subway.member.dto.MemberRequest;
 import nextstep.subway.member.dto.MemberResponse;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.net.URI;
+import java.util.Objects;
 
 @RestController
 public class MemberController {
-    private MemberService memberService;
+    private final MemberService memberService;
 
     public MemberController(MemberService memberService) {
         this.memberService = memberService;
@@ -32,20 +33,27 @@ public class MemberController {
     }
 
     @GetMapping("/members/me")
-    public ResponseEntity<MemberResponse> findMemberOfMine() {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        LoginMember loginMember = (LoginMember) authentication.getPrincipal();
-        return ResponseEntity.ok().body(memberService.findMember(loginMember.getId()));
+    public ResponseEntity<MemberResponse> findMemberOfMine(@AuthenticationPrincipal UserDetails principal) {
+        return ResponseEntity.ok().body(memberService.findMember(principal.getId()));
     }
 
     @PutMapping("/members/{id}")
-    public ResponseEntity<MemberResponse> updateMember(@PathVariable Long id, @RequestBody MemberRequest param) {
+    public ResponseEntity<MemberResponse> updateMember(@PathVariable Long id, @RequestBody MemberRequest param,
+                                                                @AuthenticationPrincipal UserDetails principal) {
+        if (!Objects.equals(id, principal.getId())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         memberService.updateMember(id, param);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/members/{id}")
-    public ResponseEntity<MemberResponse> deleteMember(@PathVariable Long id) {
+    public ResponseEntity<MemberResponse> deleteMember(@PathVariable Long id, @AuthenticationPrincipal UserDetails principal) {
+        if (!Objects.equals(id, principal.getId())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         memberService.deleteMember(id);
         return ResponseEntity.noContent().build();
     }
